📐moved and renamed form atomic design
