+++ Frontend/src/app/app.component.html
@@ -9,7 +9,7 @@
     </div>
     <app-mood-drop-down class="col-start-4 col-span-1 row-start-1 row-span-1"></app-mood-drop-down>
     <div  class="col-start-5 col-span-1 row-start-1 row-span-1" > 
-      <app-info-atom></app-info-atom>
+      <!-- <app-info-atom></app-info-atom> -->
       <app-profile-atomic></app-profile-atomic>
     </div>
     <ng-container>
